gh-2063 - reverted incorrect changes
