+++ core/store/src/main/java/uk/gov/gchq/gaffer/store/operation/handler/AddStorePropertiesToLibraryHandler.java
@@ -31,7 +31,6 @@ public class AddStorePropertiesToLibraryHandler implements OperationHandler<AddS
 
     @Override
     public Void doOperation(final AddStorePropertiesToLibrary operation, final Context context, final Store store) throws OperationException {
-        prepareOperation(operation, context, store);
         GraphLibrary graphLibrary = store.getGraphLibrary();
         if (null == graphLibrary) {
             throw new OperationException(String.format(ERROR_ADDING_STORE_TO_STORE_S, THE_STORE_DOES_NOT_HAVE_A_GRAPH_LIBRARY));
