+++ core/store/src/main/java/uk/gov/gchq/gaffer/store/operation/handler/CountGroupsHandler.java
@@ -33,7 +33,6 @@ public class CountGroupsHandler implements OutputOperationHandler<CountGroups, G
     public GroupCounts doOperation(final CountGroups operation,
                                    final Context context, final Store store)
             throws OperationException {
-        prepareOperation(operation, context, store);
         final GroupCounts groupCounts = new GroupCounts();
         try {
             int count = 0;
