+++ core/store/src/main/java/uk/gov/gchq/gaffer/store/operation/handler/GetTraitsHandler.java
@@ -36,7 +36,6 @@ public class GetTraitsHandler implements OutputOperationHandler<GetTraits, Set<S
 
     @Override
     public Set<StoreTrait> doOperation(final GetTraits operation, final Context context, final Store store) throws OperationException {
-        prepareOperation(operation, context, store);
         if (!operation.isCurrentTraits()) {
             return store.getTraits();
         }
