+++ core/store/src/main/java/uk/gov/gchq/gaffer/store/operation/handler/GetVariableHandler.java
@@ -24,7 +24,6 @@ import uk.gov.gchq.gaffer.store.Store;
 public class GetVariableHandler implements OperationHandler<GetVariable> {
     @Override
     public Object doOperation(final GetVariable operation, final Context context, final Store store) throws OperationException {
-        prepareOperation(operation, context, store);
         if (operation.getVariableName() == null) {
             throw new IllegalArgumentException("Variable name cannot be null");
         }
