+++ core/store/src/main/java/uk/gov/gchq/gaffer/store/operation/handler/OperationChainHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2019 Crown Copyright
+ * Copyright 2017-2018 Crown Copyright
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,7 +25,6 @@ import uk.gov.gchq.gaffer.store.operation.OperationChainValidator;
 import uk.gov.gchq.gaffer.store.optimiser.OperationChainOptimiser;
 import uk.gov.gchq.koryphe.ValidationResult;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -34,49 +33,36 @@ import java.util.List;
  * @param <OUT> the output type of the operation chain
  */
 public class OperationChainHandler<OUT> implements OutputOperationHandler<OperationChain<OUT>, OUT> {
-    private OperationChainValidator opValidator;
-
-    public OperationChainHandler(final OperationChainValidator opValidator) {
-        this.opValidator = opValidator;
-    }
+    private final OperationChainValidator opChainValidator;
+    private final List<OperationChainOptimiser> opChainOptimisers;
 
     @Override
-    public OUT doOperation(final OperationChain<OUT> operationChain,
-                           final Context context,
-                           final Store store) throws OperationException {
-        prepareOperation(operationChain, context, store);
+    public OUT doOperation(final OperationChain<OUT> operationChain, final Context context, final Store store) throws OperationException {
+
+        prepareOperationChain(operationChain, context, store);
 
         Object result = null;
         for (final Operation op : operationChain.getOperations()) {
             updateOperationInput(op, result);
             result = store.handleOperation(op, context);
         }
+
         return (OUT) result;
     }
 
-    @Override
-    public OperationChain<OUT> prepareOperation(final OperationChain<OUT> operation,
-                                                final Context context,
-                                                final Store store) {
-        final List<OperationChainOptimiser> opChainOptimisers =
-                new ArrayList<>();
-
-        final ValidationResult validationResult = opValidator.validate(operation, context.getUser(), store);
-
+    public <O> OperationChain<O> prepareOperationChain(final OperationChain<O> operationChain, final Context context, final Store store) {
+        final ValidationResult validationResult = opChainValidator.validate(operationChain, context
+                .getUser(), store);
         if (!validationResult.isValid()) {
             throw new IllegalArgumentException("Operation chain is invalid. " + validationResult
                     .getErrorString());
         }
 
-        OperationChain<OUT> optimisedOperationChain = operation;
+        OperationChain<O> optimisedOperationChain = operationChain;
         for (final OperationChainOptimiser opChainOptimiser : opChainOptimisers) {
             optimisedOperationChain = opChainOptimiser.optimise(optimisedOperationChain);
         }
         return optimisedOperationChain;
-    }
-
-    public OperationChainValidator getOpValidator() {
-        return opValidator;
     }
 
     protected void updateOperationInput(final Operation op, final Object result) {
@@ -95,9 +81,22 @@ public class OperationChainHandler<OUT> implements OutputOperationHandler<Operat
         }
     }
 
+    public OperationChainHandler(final OperationChainValidator opChainValidator, final List<OperationChainOptimiser> opChainOptimisers) {
+        this.opChainValidator = opChainValidator;
+        this.opChainOptimisers = opChainOptimisers;
+    }
+
     private void setOperationInput(final Operation op, final Object result) {
         if (null == ((Input) op).getInput()) {
             ((Input) op).setInput(result);
         }
+    }
+
+    protected OperationChainValidator getOpChainValidator() {
+        return opChainValidator;
+    }
+
+    protected List<OperationChainOptimiser> getOpChainOptimisers() {
+        return opChainOptimisers;
     }
 }
